Search videos from it's own live component
* Move add video icon to the player controls section at the footer
