+++ lib/livedj_web/live/sessions/room_live/list.html.heex
@@ -12,31 +12,3 @@
     topic={"room:#{@room.id}"}
   />
 </div>
-<div class="flex mt-4">
-  <.live_component
-    id={"video_search_bar_#{@room.id}"}
-    module={LivedjWeb.Components.SearchBarComponent}
-    form={@search_form}
-    search_result={@search_result}
-  >
-    <:button>
-      <.button class="
-        cursor-default w-6 h-6 !p-0 
-        align-middle ml-2 bg-zinc-300 dark:bg-zinc-700
-        transition-all duration-300
-        shadow-[3.0px_3.0px_1px_0.5px_rgba(24,24,27,0.5)]
-        hover:shadow-[2.5px_2.5px_1px_0.5px_rgba(24,24,27,0.9)]
-        active:shadow-[0.5px_0.5px_1px_0.5px_rgba(24,24,27,0.2)]
-        dark:shadow-[2.5px_2.5px_1px_0.5px_rgba(250,250,255,0.4)]
-        dark:hover:shadow-[2.0px_2.0px_1px_0.5px_rgba(250,250,255,0.6)]
-        dark:active:shadow-[0.5px_0.5px_1px_0.5px_rgba(250,250,255,0.2)]
-        hover:bg-zinc-300 dark:hover:bg-zinc-700
-        active:bg-zinc-200 dark:active:bg-zinc-800 group
-        text-zinc-900 dark:text-zinc-100
-        active:text-green-500 dark:active:text-green-500
-      ">
-        <.icon name="hero-plus" class="w-4 h-4" />
-      </.button>
-    </:button>
-  </.live_component>
-</div>
